+++ test/plausible_web/controllers/auth_controller_test.exs
@@ -1808,16 +1808,14 @@ defmodule PlausibleWeb.AuthControllerTest do
       password: password
     })
     |> recycle()
-    |> Map.put(:secret_key_base, secret_key_base())
-    |> Plug.Conn.put_req_header("x-forwarded-for", random_ip())
+    |> prepare_conn()
   end
 
   defp set_remember_2fa_cookie(conn, user) do
     conn
     |> PlausibleWeb.TwoFactor.Session.maybe_set_remember_2fa(user, "true")
     |> recycle()
-    |> Map.put(:secret_key_base, secret_key_base())
-    |> Plug.Conn.put_req_header("x-forwarded-for", random_ip())
+    |> prepare_conn()
   end
 
   defp mock_captcha_success() do
@@ -1841,11 +1839,5 @@ defmodule PlausibleWeb.AuthControllerTest do
          }}
       end
     )
-  end
-
-  defp secret_key_base() do
-    :plausible
-    |> Application.fetch_env!(PlausibleWeb.Endpoint)
-    |> Keyword.fetch!(:secret_key_base)
   end
 end