+++ test/plausible_web/live/shields/countries_test.exs
@@ -38,6 +38,17 @@ defmodule PlausibleWeb.Live.Shields.CountriesTest do
       assert text_of_attr(remove_button_2, "phx-value-rule-id") == r2.id
     end
 
+    test "lists Unknown country rule", %{conn: conn, site: site} do
+      {:ok, rule} =
+        Shields.add_country_rule(site, %{"country_code" => "ZZ"})
+
+      conn = get(conn, "/#{site.domain}/settings/shields/countries")
+      resp = html_response(conn, 200)
+
+      assert resp =~ "Unknown"
+      assert find(resp, "#remove-country-rule-#{rule.id}")
+    end
+
     test "add rule button is rendered", %{conn: conn, site: site} do
       conn = get(conn, "/#{site.domain}/settings/shields/countries")
       resp = html_response(conn, 200)
@@ -103,6 +114,22 @@ defmodule PlausibleWeb.Live.Shields.CountriesTest do
       tooltip = text_of_attr(html, "#country-#{id}", "title")
       assert tooltip =~ "Added at #{Date.utc_today()}"
       assert tooltip =~ "by #{added_by}"
+    end
+
+    test "submitting Unknown (ZZ) country saves it", %{conn: conn, site: site} do
+      lv = get_liveview(conn, site)
+
+      lv
+      |> element("form")
+      |> render_submit(%{
+        "country_rule[country_code]" => "ZZ"
+      })
+
+      html = render(lv)
+
+      assert html =~ "Unknown"
+
+      assert [%{country_code: "ZZ"}] = Shields.list_country_rules(site)
     end
 
     test "clicking Remove deletes the rule", %{conn: conn, site: site} do